+++ object_tools/options.py
@@ -1,9 +1,8 @@
+from __future__ import unicode_literals
+
 from django import forms
 from django.conf import settings
-try:
-    from django.conf.urls.defaults import patterns, url
-except ImportError:
-    from django.conf.urls import patterns, url
+from django.conf.urls import url
 from django.contrib.admin import helpers
 from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
@@ -53,7 +52,7 @@ class ObjectTool(object):
         return form
 
     def get_permission(self):
-        return u'%s_%s' % (self.name, self.model._meta.object_name.lower())
+        return '%s_%s' % (self.name, self.model._meta.object_name.lower())
 
     def has_permission(self, user):
         """
@@ -76,7 +75,7 @@ class ObjectTool(object):
         )
 
         if form:
-            for name, field in form.fields.iteritems():
+            for name, field in form.fields.items():
                 media = media + field.widget.media
 
         return media
@@ -95,16 +94,13 @@ class ObjectTool(object):
         """
         URL patterns for tool linked to _view method.
         """
-        info = (self.model._meta.app_label,)
-        # to keep backward (Django <= 1.7) compatibility
-        try:
-            info += (self.model._meta.model_name,)
-        except AttributeError:
-            info += (self.model._meta.module_name,)
-        info += (self.name,)
-        urlpatterns = patterns(
-            '', url(r'^%s/$' % self.name, self._view, name='%s_%s_%s' % info),
+        info = (
+            self.model._meta.app_label, self.model._meta.model_name,
+            self.name,
         )
+        urlpatterns = [
+            url(r'^%s/$' % self.name, self._view, name='%s_%s_%s' % info)
+        ]
         return urlpatterns
     urls = property(_urls)